backend-server/fastapi/requirements.txt
fastapi==0.115.12
uvicorn==0.33.0
sqlalchemy==1.4.15
pymysql==1.1.1
sentence-transformers==3.0.1
faiss-cpu==1.7.4
numpy>=1.24.0
pandas>=1.3.4
torch>=2.0.0
httpx>=0.24.0
langchain==0.3.27
langchain-community==0.3.16
openai>=1.0.0
langchain-openai>=0.1.0
transformers>=4.35.0
accelerate>=0.25.0
ultralytics>=8.2.0
python-multipart>=0.0.9
websockets>=12.0
